fix startup bug

+++ AntiRecall/deploy/ResumeQQStartup.cs
@@ -7,7 +7,6 @@ namespace AntiRecall.deploy
     public class ResumeQQStartup : ICommand
     {
         private static RegistryKey startupKey;
-        private static string QQ_ori_path;
 
         public bool CanExecute(object parameter)
         {
@@ -19,11 +18,17 @@ namespace AntiRecall.deploy
             startupKey = Registry.CurrentUser.OpenSubKey("Software\\Microsoft\\Windows\\CurrentVersion\\Run", true);
             string QQName = "QQ2009";
             string MyName = "AntiRecall";
+            if (-1 == Xml.antiRElement["QQPath"].IndexOf("QQ.exe"))
+            {
+                System.Windows.Forms.MessageBox.Show("未设置QQ安装目录");
+                return;
+            }
             if (IsInStartup(MyName))
             {
                 DeleteStartup(MyName);
+                CreateStartup(QQName, "\"" + Xml.antiRElement["QQPath"] + "\"  /background");
             }
-            CreateStartup(QQName, Xml.antiRElement["QQPath"]);
+            System.Windows.Forms.MessageBox.Show("已恢复QQ自启动，AntiRecall不再自启动");
         }
 
         public event EventHandler CanExecuteChanged;
@@ -35,7 +40,6 @@ namespace AntiRecall.deploy
                 Object o = startupKey.GetValue(KeyName);
                 if (o != null)
                 {
-                    QQ_ori_path = (string)o;
                     return true;
                 }
